+++ src/systems/si/units/mod.rs
@@ -6,12 +6,12 @@ mod scaled;
 
 pub use base::{Ampere, Candela, Kelvin, Kilogram, Meter, Mole, Radian, Second};
 pub use derived::{
-    CubicMeter, CubicMeterPerSecond, Gray, GrayPerSecond, Hertz, Joule, JoulePerCubicMeter,
-    JoulePerKelvin, JoulePerKilogramKelvin, JoulePerMole, JoulePerMoleKelvin, JoulePerSquareMeter,
-    KilogramPerCubicMeter, KilogramPerCubicMeterSecond, MeterPerSecond, Newton, NewtonPerMeter,
-    Pascal, PascalSecond, PerCubicMeter, PerKelvin, PerMeter, PerSecond, PerSquareKelvin, Rayl,
-    SquareMeter, SquareMeterPerKilogram, SquareMeterPerSecond, Watt, WattPerCubicMeter,
-    WattPerKilogram, WattPerMeterKelvin, WattPerSquareMeter,
+    Coulomb, CubicMeter, CubicMeterPerSecond, Farad, Gray, GrayPerSecond, Hertz, Joule,
+    JoulePerCubicMeter, JoulePerKelvin, JoulePerKilogramKelvin, JoulePerMole, JoulePerMoleKelvin,
+    JoulePerSquareMeter, KilogramPerCubicMeter, KilogramPerCubicMeterSecond, MeterPerSecond,
+    Newton, NewtonPerMeter, Pascal, PascalSecond, PerCubicMeter, PerKelvin, PerMeter, PerSecond,
+    PerSquareKelvin, Rayl, Siemens, SquareMeter, SquareMeterPerKilogram, SquareMeterPerSecond,
+    Volt, Watt, WattPerCubicMeter, WattPerKilogram, WattPerMeterKelvin, WattPerSquareMeter,
 };
 pub use scaled::{
     Centimeter, CubicMillimeter, ElectronVolt, Gram, GramPerCubicCentimeter, JoulePerMilliliter,